README
BLE Device Function Pointers Application Interface

This code is part of the larger code. The purpose of this code is to create a function pointers Application Interface enabling the BLE device to send and receive data packets to and from Smart Mobile Phone. The data packet contains various elements such as the version of current code, LED state, button states, Error code and so on.

The BLE device will then be able to use a custom API Interface commands such as BLE.WRITE and BLE READ.